+++ src/VertexHandler.ts
@@ -3,8 +3,8 @@ import { EditAggregationFunctionDialog, EditComputedAttributesFunctionDialog } f
 import { AsyncWorker } from './AsyncUtils';
 import { AttributeRenderer } from './AttributeRenderer';
 import { AttackGraphSettings } from './AttackGraphSettings';
-import { CellStyles } from './Analysis/CellStyles';
 import { NodeAttributeProvider } from './Analysis/NodeAttributeProvider';
+import { CellStyles } from './Analysis/CellStyles';
 
 
 const IMAGE_WIDTH = 24;
@@ -90,7 +90,7 @@ export const installVertexHandler = (ui: Draw.UI, worker: AsyncWorker): void => 
       dlg.init();
       void (async () => {
         if (await dlg.show()) {
-          await AttributeRenderer.refreshCellValuesUpwards(cell.cell, ui, worker);
+          await AttributeRenderer.refreshCellValuesUpwards([cell.cell], ui, worker);
           await ui.editor.graph.refresh();
         }
       })();
@@ -102,7 +102,7 @@ export const installVertexHandler = (ui: Draw.UI, worker: AsyncWorker): void => 
       dlg.init();
       void (async () => {
         if (await dlg.show()) {
-          await AttributeRenderer.refreshCellValuesUpwards(cell.cell, ui, worker);
+          await AttributeRenderer.refreshCellValuesUpwards([cell.cell], ui, worker);
           await ui.editor.graph.refresh();
         }
       })();
@@ -135,15 +135,6 @@ export const installVertexHandler = (ui: Draw.UI, worker: AsyncWorker): void => 
     if (AttackGraphSettings.isAttackGraph(ui.editor.graph) && this.graph.getSelectionCount() === 1) {
       drawFunctionHandle(this);
       drawTooltipHandle(this);
-
-      // Highlight incoming and outgoing edges
-      const cell = this.state.cell;
-      if (cell.edges) {
-        const style = new CellStyles(cell);
-        style.updateConnectedEdgesStyle(true, true);
-        ui.editor.graph.refresh();
-      }
-
       this.redrawHandles();
     }
   }
@@ -218,14 +209,6 @@ export const installVertexHandler = (ui: Draw.UI, worker: AsyncWorker): void => 
         if (functionHandle.parentNode !== null) {
           functionHandle.parentNode.removeChild(functionHandle);
         }
-      }
-
-      // Remove highlight of incoming and outgoing edges
-      const cell = this.state.cell;
-      if (cell.edges) {
-        const style = new CellStyles(cell);
-        style.updateConnectedEdgesStyle(false, false);
-        ui.editor.graph.refresh();
       }
 
       this.functionHandles = null;