+++ CHANGELOG.md
@@ -1,6 +1,22 @@
 # Changelog
 
-## [Unreleased](https://github.com/INCYDE-GmbH/drawio-plugin-attackgraphs/compare/v1.2.0...HEAD)
+## [Unreleased](https://github.com/INCYDE-GmbH/drawio-plugin-attackgraphs/compare/v1.2.2...HEAD)
+
+### Fixes and Improvements
+
+- Undos sometimes did not trigger recalculations of cell values ([#109](https://github.com/INCYDE-GmbH/drawio-plugin-attackgraphs/issues/109))
+- Enhance recalculation of cell values ([#108](https://github.com/INCYDE-GmbH/drawio-plugin-attackgraphs/issues/108))
+  - Avoid recaluclating the values of cells more than once
+  - Allow to specify several cells to simultaneously update within one function call
+
+## [1.2.2](https://github.com/INCYDE-GmbH/drawio-plugin-attackgraphs/compare/v1.2.1...v1.2.2) - 2023-03-20
+
+### Fixes and Improvements
+
+- Reconnecting nodes now triggers a recalculation of values for all affected cells ([#98](https://github.com/INCYDE-GmbH/drawio-plugin-attackgraphs/issues/98))
+- Duplicates are not shown anymore when resizing nodes ([#79](https://github.com/INCYDE-GmbH/drawio-plugin-attackgraphs/issues/79))
+
+## [1.2.1](https://github.com/INCYDE-GmbH/drawio-plugin-attackgraphs/compare/v1.2.0...v1.2.1) - 2023-01-23
 
 ### Added
 
@@ -9,6 +25,7 @@
 ### Fixes and Improvements
 
 - The width of attack graph node batches grows depending on the width of the text displayed in it
+- Opening an attack graph caused an error when using draw.io with version >= 20.6.0
 
 ## [1.2.0](https://github.com/INCYDE-GmbH/drawio-plugin-attackgraphs/compare/v1.1.0...v1.2.0) - 2022-11-11